Knowledge Base/Knowledge Base/requirements.txt
keybert
supabase
boto3
langchain
langchain-community
langchain-aws
sentence-transformers
gdown
python-dotenv
